Major update of the metrics-collector
The resource provider now has a single task that connects to Redis and the resource instances send commands via channels, following a single receiver multiple senders pattern.

The schema of the metrics-collector's Redis output has been updated, which required an update of the edgeless_benchmark functions, and the executable, too.

The metrics-collector in the  can be reset (with a configurable warm-up period following the reset) by sending a command to it via a resource instance. The single_trigger function has been added for this purpose.