Add collection of performance samples from nodes via keep-alive responses
API change: the response to a keep-alive poll is not anymore a HealthStatus but a new message called KeepAliveResponse, which includes a HealthStatus (renamed as NodeHealthStatus for naming consistency) and NodePerformanceSamples

Add a new command-line option to proxy_cli, dump performance, which saves to files the content of all the performance:* keys in Redis.

The KeepAlive section of the main loop of the orchestrator has been simplified by removing HealthStatus info from the nodes' descriptors, since this piece of info is ephemeral and only used to update the proxy in the same procedure.